Add DashCam style recording for gif format.
In this version have added the option to retains just the last specified time of recording and ignore the inital recording when saving in gif format. This option is useful when recording hard to find cases and you just want to retain the last few seconds or recording with the case being simulated.

In this have created a base class for encoder and from that class derived the gif_encoder and dashcam_encoder. gif_encoder remains mostly the same but with functions being transferred to the base class whereas in dashcam_encoder we are saving the scrren differences for the last desired time as memory bitmap and discard the older ones and recreate a complete gif file when the user closes the recording.

This option is only available when saving as gif file and on windows system.